services/market-mock/requirements.txt
mysql-connector-python
psycopg2-binary
click
oracledb
pymssql
